+++ src/widget/mod.rs
@@ -5,8 +5,8 @@
 pub mod button;
 pub mod dropdown;
 pub mod label;
-pub mod multiline_input;
 pub mod node;
 pub mod pin;
 pub mod pin_group;
 pub mod slider;
+pub mod text_box;